==> data_wrangling/02_time_conversion.ipynb <==
# Convert Unix Time into Human-Readable and Timezone-Adjusted Time

This code is particularly helpful if you are working with time in your data, and you want to be able to adjust times to specific timezones, extract certain components of time (i.e. year, hour, or day), or convert Unix time into a format that is more readable.

The time given to us in our datasets was in Unix format, which displays the time as the number of seconds that has passed since January 1st, 1970. This format is not very helpful for the visualizations that we wanted to make, so we wanted to convert the Unix Time into human-readable time first, and adjust the time based on timezone. We used this code to create our stacked bar chart, where we got the relative frequencies of number of posts by hour of day for each subreddit cluster by using this for loop on our combined Reddit post dataset to get the timezone-adjusted hour of day for each post.

For the purposes of this data manipulation demo, because of data privacy, we've subsetted 100 data observations from the long/lat geographic dataset with fake usernames to replace the 'author' column, which is what we'll be using in this data demo and in our visualizations.

## Import Libraries

Here we are importing the necessary libraries to run the conversion code. 

import os
import numpy as np
import pandas as pd
import pytz
from datetime import datetime, timedelta, date 

## Read in Data

Read in the data that you would like to perform the following data manipulations on.

The `os.getcwd()` method gets the current working directory that you are in, which should be inside the `data_wrangling` folder. However, to access the data file, we need to replace the current working directory with the directory that leads to the file. Once that has been done, we can go head with reading in the data and performing the necessary data manipulations.

DATA_DIR = os.getcwd()
DATA_DIR = DATA_DIR.replace('data_wrangling', 'synthetic_data')

location_data = pd.read_parquet(DATA_DIR + '/geo_known_synthetic.parquet')

## Convert Unix Time to Human-Readable Time

We are using the `.to_datetime()` function to convert the Unix Time into human-readable time.

location_data['converted_time'] = pd.to_datetime(location_data['created_utc'], unit='s')

## Adjust Time by Timezone + Extract Time Components

The long/lat geographical dataset contains timezone information in the following format: "Country/Major City". We used the following for-loop iteration to go through our dataset and convert the Unix Time into human-readable time that is properly adjusted based on the given timezone. We can also specifically extract out parts of the adjusted time, including date, month, day, hour, minute, seconds, and the abbreviated timezone.

for index, row in location_data.iterrows():
    tz = pytz.timezone(row['timezone'])
    ldt = datetime.fromtimestamp(row['created_utc'], tz)
    location_data.loc[index, "time_adj"] = ldt.strftime('%Y-%m-%d %H:%M:%S %Z%z').upper()
    location_data.loc[index, "date"] = ldt.strftime('%Y-%m-%d').upper()
    location_data.loc[index, "month"] = ldt.strftime('%m').upper()
    location_data.loc[index, "day"] = ldt.strftime('%d').upper()
    location_data.loc[index, "hour"] = ldt.strftime('%H').upper()
    location_data.loc[index, "tz_abbrev"] = ldt.strftime('%Z').upper()

location_data
