==> data_wrangling/04_extract_subsets.ipynb <==
# Subsetting Parts of a Dataset

You can extract certain parts of a dataset and store those subsets into separate dataframes for later use. This can be particularly helpful for when you are doing data manipulations on a very large dataset.

When we wanted to adjust the posting time by timezone for each post in our combined dataset, we ran the time conversion code on different sections of the dataset, dividing it up by subreddit cluster name, and then merged the subsets back together. Depending on how large your dataset is, you'd want to allocate more CPU cores to expedite the data manipulations process. 

For data privacy reasons, we've subsetted from our given datasets, created fake usernames to replace the 'author' column in some of our datasets, and made a combined file in our previous demo that we will be doing manipulations on in this demo.

## Import Libraries

import os
import pandas as pd
import pytz
from datetime import datetime
import warnings

## Read in Dataset

This is the dataset that we got from the previous data manipulation demo where we combined sections of multiple datasets.

The `os.getcwd()` method gets the current working directory that you are in, which should be inside the `data_wrangling` folder. However, to access the data file, we need to replace the current working directory with the directory that leads to the file. Once that has been done, we can go head with reading in the data and performing the necessary data manipulations.

DATA_DIR = os.getcwd()
DATA_DIR = DATA_DIR.replace('data_wrangling', 'synthetic_data')

combined_df = pd.read_parquet(DATA_DIR + '/merged_combined_file.parquet')

combined_df

## Extract Subsets of Data

Using the `.isin()` method, we can extract out a select portion of the dataset based on dividing the dataset up by a certain variable.

opts_0 = ['Anime', 'Canada', 'Cars', 'Cryptocurrencies', 'Gaming 1', 'Gaming 2']
subset_0 = combined_df.loc[combined_df['cluster_name'].isin(opts_0)]

opts_1 = ['Gaming 3', 'Gaming 4', 'Gaming 5', 'General interest 1', 'General interest 2', 'General interest 3']
subset_1 = combined_df.loc[combined_df['cluster_name'].isin(opts_1)]

opts_2 = ['General interest 4', 'General interest 5', 'General interest 6', 'General interest 7', 'History', 'Hobbies 1']
subset_2 = combined_df.loc[combined_df['cluster_name'].isin(opts_2)]

opts_3 = ['Hobbies 2', 'LGBT', 'Movies and TV', 'Music', 'Personal matters', 'Politics']
subset_3 = combined_df.loc[combined_df['cluster_name'].isin(opts_3)]

opts_4 = ['Pornography', 'Programming', 'Recreational drugs', 'Rest of world', 'Self-improvement', 'USA'] 
subset_4 = combined_df.loc[combined_df['cluster_name'].isin(opts_4)]

## Data Manipulations

After getting a subset of the data, you can run your desired data manipulations on that section of the data. Repeat this process for the rest of the data, making sure to change the `opts` and the name of the subsetted dataframe each time you run the data manipulation. 

warnings.filterwarnings("ignore")

for index, row in subset_0.iterrows():
    tz = pytz.timezone(row['timezone'])
    ldt = datetime.fromtimestamp(row['created_utc'], tz)
    subset_0.loc[index, "date"] = ldt.strftime('%Y-%m-%d').upper()
    subset_0.loc[index, "hour"] = ldt.strftime('%H').upper()
    subset_0.loc[index, "day_of_week"] = pd.Timestamp(subset_0.loc[index, "date"]).day_name()

for index, row in subset_1.iterrows():
    tz = pytz.timezone(row['timezone'])
    ldt = datetime.fromtimestamp(row['created_utc'], tz)
    subset_1.loc[index, "date"] = ldt.strftime('%Y-%m-%d').upper()
    subset_1.loc[index, "hour"] = ldt.strftime('%H').upper()
    subset_1.loc[index, "day_of_week"] = pd.Timestamp(subset_1.loc[index, "date"]).day_name()

for index, row in subset_2.iterrows():
    tz = pytz.timezone(row['timezone'])
    ldt = datetime.fromtimestamp(row['created_utc'], tz)
    subset_2.loc[index, "date"] = ldt.strftime('%Y-%m-%d').upper()
    subset_2.loc[index, "hour"] = ldt.strftime('%H').upper()
    subset_2.loc[index, "day_of_week"] = pd.Timestamp(subset_2.loc[index, "date"]).day_name()

for index, row in subset_3.iterrows():
    tz = pytz.timezone(row['timezone'])
    ldt = datetime.fromtimestamp(row['created_utc'], tz)
    subset_3.loc[index, "date"] = ldt.strftime('%Y-%m-%d').upper()
    subset_3.loc[index, "hour"] = ldt.strftime('%H').upper()
    subset_3.loc[index, "day_of_week"] = pd.Timestamp(subset_3.loc[index, "date"]).day_name()

for index, row in subset_4.iterrows():
    tz = pytz.timezone(row['timezone'])
    ldt = datetime.fromtimestamp(row['created_utc'], tz)
    subset_4.loc[index, "date"] = ldt.strftime('%Y-%m-%d').upper()
    subset_4.loc[index, "hour"] = ldt.strftime('%H').upper()
    subset_4.loc[index, "day_of_week"] = pd.Timestamp(subset_4.loc[index, "date"]).day_name()

## Merge Subsets

Once all of the data manipulations are complete, merge all of the subsets together to get the completed dataset.

subset_01 = subset_0.merge(subset_1, how='outer')

subset_23 = subset_2.merge(subset_3, how='outer')

subset_234 = subset_23.merge(subset_4, how='outer')

fin_comb = subset_01.merge(subset_234, how='outer')

fin_comb

## Export to Parquet File

You can save your data manipulation changes as a `.parquet` file, so whenever you want to use this data, you don't need to run all of the data manipulations again. Parquet is a type of file format that makes loading in and saving datasets much more efficient. 

fin_comb.to_parquet(DATA_DIR + '/final_combinedsubclus.parquet')
